Changed the type Preferences::index_size
index_size is maximum size of indexfile
The whole indexfile has to fit in memory,
the constraint is that all arrays in memory should
be indexable with the help of signed int.
Use SPDX-License tag.

+++ cinelerra/interfaceprefs.C
@@ -1,23 +1,7 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
 
-/*
- * CINELERRA
- * Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
- * 
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- * 
- */
+// This file is a part of Cinelerra-CVE
+// Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
 
 #include "bcbar.h"
 #include "bcresources.h"
@@ -155,7 +139,7 @@ void InterfacePrefs::show()
 		1));
 
 // Index file size
-	sprintf(string, "%" PRId64, pwindow->thread->preferences->index_size);
+	sprintf(string, "%d", pwindow->thread->preferences->index_size);
 	add_subwindow(isize = new IndexSize(maxw, ybix[1], pwindow, string));
 
 // Number of index files to keep
@@ -270,9 +254,9 @@ IndexSize::IndexSize(int x,
 
 int IndexSize::handle_event()
 {
-	long result;
+	int result;
 
-	result = atol(get_text());
+	result = atoi(get_text());
 	if(result < 64000) result = 64000;
 	pwindow->thread->preferences->index_size = result;
 	return 1;
